% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_fdiv.R
\name{fd_fdiv}
\alias{fd_fdiv}
\title{Compute Functional Divergence FDiv}
\usage{
fd_fdiv(traits)
}
\arguments{
\item{traits}{The matrix dataset for which you want to compute the index}
}
\value{
The value of FDiv (numeric of length 1)
}
\description{
Compute Functional Divergence FDiv
}
\examples{
data(traits_birds)
fd_fdiv(traits_birds)

}
\references{
Villéger S., Mason N. W. H., Mouillot D. (2008), New multidimensional
functional diversity indices for a multifaceted framework in functional
ecology, Ecology 89(8), \doi{10.1890/07-1206.1}
}
