% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_fric_intersect.R
\name{fd_fric_intersect}
\alias{fd_fric_intersect}
\title{Intersection between functional volume of pairs of sites}
\usage{
fd_fric_intersect(traits, sp_com, stand = FALSE)
}
\arguments{
\item{traits}{The matrix dataset for which you want to compute the index}

\item{sp_com}{Site-species matrix with sites as rows and species as columns
if not provided, the function considers all species with equal
abundance in a single site. This can be either a \code{matrix},
a \code{data.frame}, or a \code{\link[Matrix:Matrix]{Matrix::Matrix()}} object.}

\item{stand}{a boolean indicating whether to standardize FRic values over
the observed maximum over all species (default: \code{FALSE}).
This scales FRic between 0 and 1.
\strong{NB}: The maximum FRic values only considers species that are
present in \strong{both} site-species and trait matrices.
If you want to consider species that are absent
in the site-species matrix, add corresponding columns of 0s.}
}
\value{
a data.frame with three columns:
\itemize{
\item \code{first_site} the names of the first site used in the pair \code{sp_com},
\item \code{second_site} the names of the first site used in the pair,
\item \code{FRic_intersect} the values of intersection of functional volumes of each
pair of site.
}
}
\description{
Compute volume of the intersection of functional volumes of all pairs
of sites (including self-intersection, which corresponds to their functional
volumes). Note that when standardizing functional volumes of intersections,
this function uses the functional volume all provided traits,
thus standardized self-intersection volumes can be lower than one.
}
\examples{
data(traits_birds)
fd_fric_intersect(traits_birds)

}
