% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_raoq.R
\name{fd_raoq}
\alias{fd_raoq}
\title{Compute Rao's entropy index (Rao's Q)}
\usage{
fd_raoq(traits, sp_com)
}
\arguments{
\item{traits}{The matrix dataset for which you want to compute the index}

\item{sp_com}{Abundance matrix}
}
\value{
The value of Rao's Q (numeric of length 1)
}
\description{
Compute Rao's entropy index (Rao's Q)
}
\examples{
data(traits_birds)
fd_raoq(traits_birds)

}
