% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_memoise.R
\name{fundiversity-options}
\alias{fundiversity-options}
\title{Options for \pkg{fundiversity}}
\description{
The memoisation is the convex hull computation in \pkg{fundiversity} is
controlled via the \code{fundiversity.memoise} option:
\itemize{
\item if unset, the default is to use memoisation if \pkg{memoise} was installed
when \pkg{fundiversity} was loaded, and not to use memoisation otherwise.
\item if \code{options(fundiversity.memoise = TRUE)}, memoisation is used and an error
is thrown if \pkg{memoise} is not installed.
\item if \code{options(fundiversity.memoise = FALSE)}, memoisation is not used.
}
}
