% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundiversity-package.R
\docType{package}
\name{fundiversity-package}
\alias{fundiversity}
\alias{fundiversity-package}
\title{fundiversity: Easy Computation of Functional Diversity Indices}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Computes six functional diversity indices. These are namely, Functional Divergence (FDiv), Function Evenness (FEve), Functional Richness (FRic), Functional Richness intersections (FRic_intersect), Functional Dispersion (FDis), and Rao's entropy (Q) (reviewed in Villéger et al. 2008 \doi{10.1890/07-1206.1}). Provides efficient, modular, and parallel functions to compute functional diversity indices (preprint: \doi{10.32942/osf.io/dg7hw}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://funecology.github.io/fundiversity/}
  \item \url{https://github.com/funecology/fundiversity}
  \item Report bugs at \url{https://github.com/funecology/fundiversity/issues}
}

}
\author{
\strong{Maintainer}: Matthias Grenié \email{matthias.grenie@gmail.com} (\href{https://orcid.org/0000-0002-4659-7522}{ORCID})

Authors:
\itemize{
  \item Hugo Gruson (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
}

}
\keyword{internal}
