% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_species_without_trait.R
\name{remove_species_without_trait}
\alias{remove_species_without_trait}
\title{Remove species with missing trait data}
\usage{
remove_species_without_trait(trait_mat)
}
\arguments{
\item{trait_mat}{a matrix of trait values with species as rows and traits as
columns}
}
\description{
Remove species with missing trait data
}
