% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_sp_birds}
\alias{site_sp_birds}
\title{Site-species matrix of birds along a Tropical Gradient}
\format{
A matrix with 217 columns (1 per species) and 8 rows:
\describe{
\item{elev_250}{is species present at 250 m elevation? 0=No, 1=Yes}
\item{elev_500}{is species present at 500 m elevation? 0=No, 1=Yes}
\item{elev_1000}{is species present at 1000 m elevation? 0=No, 1=Yes}
\item{elev_1500}{is species present at 1500 m elevation? 0=No, 1=Yes}
\item{elev_2000}{is species present at 2000 m elevation? 0=No, 1=Yes}
\item{elev_2500}{is species present at 2500 m elevation? 0=No, 1=Yes}
\item{elev_3000}{is species present at 3000 m elevation? 0=No, 1=Yes}
\item{elev_3500}{is species present at 3500 m elevation? 0=No, 1=Yes}
}
}
\source{
Nowak, Larissa et al. (2019), Data from: Projecting consequences of
global warming for the functional diversity of fleshy-fruited plants and
frugivorous birds along a tropical elevational gradient, Dryad, Dataset,
\doi{10.5061/dryad.c0n737b}
}
\usage{
site_sp_birds
}
\description{
Presences and absences of birds at different elevations along a tropical
gradient. Species names are indicated as column names.
}
\keyword{datasets}
