% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_in_common.R
\name{species_in_common}
\alias{species_in_common}
\title{List of species in common between trait dataset and site-species matrix}
\usage{
species_in_common(traits, site_sp)
}
\arguments{
\item{traits}{a \code{matrix} or \code{data.frame} that describe the traits of species
with each species as a row and traits as columns}

\item{site_sp}{a \code{matrix}, \code{data.frame} or a \code{Matrix} with the
abundances or occurrences of species in each site,
with sites as rows and species as columns}
}
\description{
This functions returns the list of species in common between trait dataset
and site-species matrix. If no species are found in common, it errors.
When the species number is lower in either the trait dataset or
the site-species matrix it shows a message.
}
