% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits_birds}
\alias{traits_birds}
\title{Functional Traits of Frugivorous Birds along a Tropical Gradient}
\format{
A data frame with 217 rows and 4 variables:
\describe{
\item{Bill.width..mm.}{bill width, in mm}
\item{Bill.length..mm.}{bill length, in mm}
\item{Kipp.s.index}{Kipp's index indicating wing Pointedness}
\item{Bodymass..g.}{adult's bodymass, in g}
}
}
\source{
Nowak, Larissa et al. (2019), Data from: Projecting consequences of
global warming for the functional diversity of fleshy-fruited plants and
frugivorous birds along a tropical elevational gradient, Dryad, Dataset,
\doi{10.5061/dryad.c0n737b}
}
\usage{
traits_birds
}
\description{
A dataset containing some functional traits of frugivorous birds in the Manú
biosphere reserve, southeast Peru. Given are species mean trait values.
The row names of the dataset give species names. Morphological
traits have been measured on museum specimen following Eck et al.(2011).
Traits have been measured only for adult and, if possible, for a minimum of
two female and two male specimens. Body mass was taken
from Dunning et al. (2007).
}
\keyword{datasets}
