% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits_plants}
\alias{traits_plants}
\title{Functional Traits of Fleshy-fruit plants along a Tropical Gradient}
\format{
A data frame with 392 rows and 4 variables:
\describe{
\item{Fruit.length..mm.}{fruit length, in mm}
\item{Fruit.width..mm.}{fruit width, in mm}
\item{Plant.height..m.}{plant height, in m}
\item{Crop.mass..g.}{seed mass, in g}
}
}
\source{
Nowak, Larissa et al. (2019), Data from: Projecting consequences of
global warming for the functional diversity of fleshy-fruited plants and
frugivorous birds along a tropical elevational gradient, Dryad, Dataset,
\doi{10.5061/dryad.c0n737b}
}
\usage{
traits_plants
}
\description{
Taxonomy and functional traits of 392 fleshy-fruited plant species from
the Manu National Park in south-east Peru. Given are fruit length and
width (mm), plant height (m) and crop mass (g). Fruit traits have been
measured on fresh fruit samples. Number of fruits per plant
(used to determine the crop mass) and plant height have been estimated
in the field. Given are species mean trait values, except for 23 species
for which no species-level data was available and genus mean trait values
are given (see column "Genus mean"). Species names are indicated as
row names.  Plant.height..m. Crop.mass..g.
}
\keyword{datasets}
