% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_fdiv.R
\name{fd_fdiv}
\alias{fd_fdiv}
\title{Compute Functional Diversity FDiv}
\usage{
fd_fdiv(data)
}
\arguments{
\item{data}{The matrix dataset for which you want to compute the index}
}
\value{
The value of FDiv (numeric of length 1)
}
\description{
Compute Functional Diversity FDiv
}
\examples{
data(traits_birds)
fd_fdiv(traits_birds)

}
