% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_fric.R
\name{fd_fric}
\alias{fd_fric}
\title{Compute Functional Richness FRic}
\usage{
fd_fric(data)
}
\arguments{
\item{data}{The matrix dataset for which you want to compute the index}
}
\value{
The value of FRic (numeric of length 1)
}
\description{
Compute Functional Richness FRic
}
\examples{
data(traits_birds)
fd_fric(traits_birds)

}
\references{
Cornwell W. K., Schwilk D. W., Ackerly D. D. (2006), A trait-based test for
habitat filtering; convex hull volume, Ecology 84(6),
\doi{0012-9658(2006)87[1465:ATTFHF]2.0.CO;2}
}
