% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_fric.R
\name{fd_fric}
\alias{fd_fric}
\title{Compute Functional Richness FRikc}
\usage{
fd_fric(data)
}
\arguments{
\item{data}{The matrix dataset for which you want to compute the index}
}
\description{
Compute Functional Richness FRikc
}
\examples{
data(traits_birds)
fd_fric(traits_birds)

}
