% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_fric.R
\name{fd_fric}
\alias{fd_fric}
\title{Compute Functional Richness (FRic)}
\usage{
fd_fric(traits, sp_com)
}
\arguments{
\item{traits}{The matrix dataset for which you want to compute the index}

\item{sp_com}{Site-species matrix with sites as rows and species as columns
if not provided, the function considers all species with equal
abundance in a single site}
}
\value{
The value of FRic (numeric of length 1)
}
\description{
Functional Richness is computed as the volume of the convex hull from all
included traits.
}
\examples{
data(traits_birds)
fd_fric(traits_birds)

}
\references{
Cornwell W. K., Schwilk D. W., Ackerly D. D. (2006), A trait-based test for
habitat filtering; convex hull volume, Ecology 84(6),
\doi{10.1890/0012-9658(2006)87[1465:ATTFHF]2.0.CO;2}
}
